+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added support for switch valves
+- Added diagnostic sensors "Compressor failure", "Source actuator failure", "Source failure", "Source pressure failure",
+  "Three-phase failure" and "VDF failure" to the heat pump device
 
 ### Fixed